updated makefile
