+++ .gitignore
@@ -3,3 +3,4 @@
 /**/*.img
 /bochsout.txt
 /bx_enh_dbg.ini
+.vscode
